MdeModulePkg PeiCore: Add PEIM migration support
This change enhances existing pointer conversion and code relocation mechanisms in the PEI core such that modules are migrated by the core from CAR to permanent memory and related function pointers such as those in the PPI list that point to those migrated ranges are updated to the permanent memory addresses.

On block boot media such as eMMC and UFS, the BIOS code is not MMIO accessible as opposed to SPI flash and may be executed from CAR. This change moves installed code during the permanent memory transition prior to cache flush.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Michael Kubacki <michael.a.kubacki@intel.com>