MdeModulePkg/Variable: RT UEFI Variable Cache Support
This change reduces SMIs for GetVariable () by maintaining a
UEFI variable cache in Runtime DXE in addition to the pre-
existing cache in SMRAM. When the Runtime Service GetVariable()
is invoked the Runtime DXE cache is first checked before
triggering an SMI to VariableSmm. This can improve overall
system performance by servicing variable read requests without
rendezvousing all cores into SMM.

The following points should be kept in mind:

1. Only data previously read or written will be in the runtime
cache. This means GetVariable () requests for a variable that
does not exist will trigger an SMI to attempt to fetch the data
from non-volatile storage.

2. All SetVariable () requests will trigger an SMI. This occurs
regardless of whether the variable is volatile or non-volatile.

3. Both volatile and non-volatile variables are cached in a runtime
buffer. To reflect the design currently in place in the EDK II
driver, they are cached in separate buffers.

4. The cache in Runtime DXE and SMM are intended to be exact copies
of one another. All SMM variable accesses only return data from the
SMM cache. The runtime caches are only updated after the variable I/O
operation is successful in SMM. The runtime caches are only updated
from SMM.

5. Synchronization mechanisms are in place to ensure the runtime cache
content integrity with the SMM cache. These may result in updates to
runtime cache that are the same in content but different in offset and
size from updates to the SMM cache.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Michael Kubacki <michael.a.kubacki@intel.com>