MdeModulePkg PeiCore: Add the ConvertPointer service (CVE-2019-11098)
REF:https://bugzilla.tianocore.org/show_bug.cgi?id=1614

This patch introduces a new pointer conversion service for PEI. The
service interface is provided by a new PPI added to MdeModulePkg
called PEI_CONVERT_POINTER_PPI. The PPI is published by the PEI
Foundation before Temporary RAM is disabled (before the PEI
Foundation invokes TemporaryRamDone ()).

The PPI provides a single function "ConvertPointer ()" to convert a
pointer to an address in a pre-memory firmware volume that was
migrated by the core to the corresponding address in permanent
memory. Note that although the heap and stack in Temporary RAM are
migrated by the PEI Foundation addresses in those Temporary RAM
address ranges are not allowed to be converted as the state of the
heap and stack at the time the service is invoked is unknown
relative to the state of the object pointed to in Temporary RAM.

This PPI is currently not published in the PI Specification.

Cc: Dandan Bi <dandan.bi@intel.com>
Cc: Liming Gao <liming.gao@intel.com>
Cc: Jian J Wang <jian.j.wang@intel.com>
Cc: Hao A Wu <hao.a.wu@intel.com>
Signed-off-by: Michael Kubacki <michael.a.kubacki@intel.com>