UefiCpuPkg/SecMigrationPei: Add initial PEIM (CVE-2019-11098)
REF:https://bugzilla.tianocore.org/show_bug.cgi?id=1614

Adds a new PEIM (SecMigrationPei) that republishes structures
produced in SEC such as PPIs passed to the PEI Foundation via
the PpiList provided to PeiCore ().

A platform can optionally include this module based on whether
this behavior is needed by the platform's requirements.

This is done because SEC modules might not be shadowed in some
platforms due to space constraints or special alignment
requirements. The SecMigrationPei module locates interfaces that
are SEC-hosted and reinstalls the interface from a memory buffer
allocated in permanent memory.

This is important if pre-memory address access is forbidden after
memory initialization. An example is if all pre-memory code is
executed from CPU cache and system security requirements do not
allow SPI bus fetches are the pre-memory code is verified in cache.

Cc: Eric Dong <eric.dong@intel.com>
Cc: Ray Ni <ray.ni@intel.com>
Cc: Laszlo Ersek <lersek@redhat.com>
Signed-off-by: Michael Kubacki <michael.a.kubacki@intel.com>