UefiCpuPkg/SecCore: Add GDT migration (CVE-2019-11098)
Migrates the GDT in SecTemporaryRamDone (). The migration process
involves copying the currently loaded GDT into a newly allocated
memory buffer and loading the GDT from the memory buffer so the
GDT is located in permanent memory.

This is done because when Temporary RAM is disabled (e.g. CAR), the
cache will be flushed. Some systems have security requirements that
prohibit fetching content from SPI flash (or other non-volatile
storage) after the content is verified in the CPU cache.

To reduce the overall flow complexity the GDT migration is not
performed conditionally. Loading the GDT from a newly allocated
memory is not considered harmful for systems that do not have a
similar security requirement.

Cc: Eric Dong <eric.dong@intel.com>
Cc: Laszlo Ersek <lersek@redhat.com>
Cc: Ray Ni <ray.ni@intel.com>
Signed-off-by: Michael Kubacki <michael.a.kubacki@intel.com>