IntelFsp2WrapperPkg/PeiFspWrapperSecMigrationLibSample: Add initial library (CVE-2019-11098)
REF:https://bugzilla.tianocore.org/show_bug.cgi?id=1614

Adds a new library to migrate SEC structures produced in the FSP
wrapper SEC modules.

A platform can optionally choose to use this library if migrating
SEC-hosted interfaces is required.

This is important if pre-memory address access is forbidden after
memory initialization. An example is if all pre-memory code is
executed from CPU cache and system security requirements do not
allow SPI bus fetches are the pre-memory code is verified in cache.

Cc: Chasel Chiu <chasel.chiu@intel.com>
Cc: Nate DeSimone <nathaniel.l.desimone@intel.com>
Cc: Star Zeng <star.zeng@intel.com>
Signed-off-by: Michael Kubacki <michael.a.kubacki@intel.com>