.github/dependabot.yml: Enable dependabot
Enables dependabot in this repo so maintainers can better alerted
when dependency updates are available.

This GitHub action will automatically create pull requests and
summarize the dependency details. Because it is a pull request,
the CI system will validate the dependency update in the pull
request.

Configures dependabot for:

1. PIP module updates
2. Submodule updates
3. GitHub action updates

The maintainers/reviewers of the .github directory were added as
pull request reviewers so they can be notified when the pull request
is available.

Note that the `do-not-merge-pr` label is added to all pull requests
created by dependabot as the TianoCore project prefers individual
developers to send patches to the mailing list rather than use the
dependabot pull request directly. So these pull requests are used
for alerting, CI checks, and reference only.

Cc: Sean Brogan <sean.brogan@microsoft.com>
Cc: Michael D Kinney <michael.d.kinney@intel.com>
Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>