.github/workflows: Add PR formatting validator
This workflow runs when GitHub PRs are modified (edited, opened,
reopened, and synchronized) to perform basic validation of the PR
title and description.

Right now, this includes:

- Checking that PR template text does not remain in the PR description.
- Checking that the PR body is not empty.

If a check fails, a GitHub comment will be left on the PR and a
PR status check failure will be present on the PR until the issue
is resolved.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>